refactor: move mock Data Path API into denokv_testing
We need to use it in both test_kv and test_datapath.
